backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
matplotlib==3.9.0
tqdm==4.66.4
gunicorn==21.2.0
aiohttp==3.10.0
requests==2.32.3
supabase==2.3.4
python-dotenv==1.0.0